+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Hover actions
 - Command for loading files into the repl.
 - Ability to paste multiple valid Haskell lines into the repl.
 - `repl.paste_info` and `repl.cword_info` functions.